Create Product TERMINADO

+++ src/category/category.controller.ts
@@ -5,15 +5,18 @@ import { Category } from './category.entity.js'
 const em = orm.em
 
 function sanitizeCategoryInput(req: Request, res: Response, next: NextFunction){
-    if(!req.body){}else{
+    if(req.body && req.body.name){
         req.body.sanitizedInput = {
             name: req.body.name,
         }
-    next()
+        next();
+    }else{
+        res.status(400).json({message: 'Invalid input'})
     }
 }
 
 async function findAll(req: Request, res: Response) {
+    const em = orm.em.fork();
     try {
         const categories = await em.find(Category, {})
         res.status(200).json({ message: 'Found all categories', data: categories })
@@ -23,9 +26,13 @@ async function findAll(req: Request, res: Response) {
 }
 
 async function findOne(req: Request, res: Response) {
+    const em = orm.em.fork();
     try {
         const id = Number.parseInt(req.params.id)
         const category = await em.findOneOrFail(Category, { id })
+        if (!category) {
+            return res.status(404).json({ message: 'Category not found' });
+        }
         res.status(200).json({ message: 'Found category', data: category })
     } catch (error: any) {
         res.status(500).json({ message: error.message })
@@ -34,6 +41,7 @@ async function findOne(req: Request, res: Response) {
 
 
 async function add(req: Request, res: Response) {
+    const em = orm.em.fork();
     try {
         req.body.sanitizedInput.id = undefined
         const category = em.create(Category, req.body.sanitizedInput)
@@ -45,9 +53,13 @@ async function add(req: Request, res: Response) {
 }
 
 async function update(req: Request, res: Response) {
+    const em = orm.em.fork();
     try {
         const id = Number.parseInt(req.params.id)
         const categoryToUpdate = await em.findOneOrFail(Category, { id })
+        if (!categoryToUpdate) {
+            return res.status(404).json({ message: 'Category not found' });
+        }
         req.body.sanitizedInput.id = id
         em.assign(categoryToUpdate, req.body.sanitizedInput)
         await em.flush()
@@ -58,9 +70,13 @@ async function update(req: Request, res: Response) {
 }
 
 async function remove(req: Request, res: Response) {
+    const em = orm.em.fork();
     try {
         const id = Number.parseInt(req.params.id)
         const category = em.getReference(Category, id)
+        if (!category) {
+            return res.status(404).json({ message: 'Category not found' });
+        }
         await em.removeAndFlush(category)
         res.status(200).send('Category removed')
     } catch (error: any) {
